+++ mcpath/facades/__init__.py
@@ -1,6 +1,21 @@
-__all__ = ["Bedrock", "Java", "Preview", "Education"]
+__all__ = [
+    "Bedrock",
+    "Java",
+    "Preview",
+    "Education",
+    "BedrockProtocol",
+    "PreviewProtocol",
+    "JavaProtocol",
+    "EducationProtocol",
+]
 
 from mcpath.facades.bedrock import Bedrock
 from mcpath.facades.preview import Preview
 from mcpath.facades.education import Education
 from mcpath.facades.java import Java
+from mcpath.facades.types import (
+    BedrockProtocol,
+    PreviewProtocol,
+    JavaProtocol,
+    EducationProtocol,
+)