+++ mcpath/platforms/win/bedrock.py
@@ -5,7 +5,7 @@ Windows Bedrock Edition
 from typing import Optional
 from os import path
 from mcpath.facades import Bedrock
-from mcpath.utils import step_back
+from mcpath.utils import step_back, get_bedrock_gdk
 
 
 class WinBedrockEdition(Bedrock):
@@ -13,35 +13,22 @@ class WinBedrockEdition(Bedrock):
     def is_old(self, path: str) -> bool:
         return "Microsoft.MinecraftUWP" in path
 
-    def _get_game_gdk(self) -> Optional[str]:
-        """
-        New path
-        """
-        p = path.expandvars(
-            "%appdata%\\Minecraft Bedrock\\users\\shared\\games\\com.mojang"
-        )
-        if path.isdir(p):
-            return p
-        return None
-
-    def _get_game_uwp(self) -> Optional[str]:
+    def _get_game_uwp(self, *paths: str) -> Optional[str]:
         """
         Old path
         """
         p = path.expandvars(
             "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftUWP_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
         )
+        p = path.join(p, *paths)
         if path.isdir(p):
             return p
         return None
 
-    def _get_game_dir(self):
-        gdk = self._get_game_gdk()
-        if gdk:
-            return gdk
-
-        # Fallback to old path
-        return self._get_game_uwp()
+    def _get_game_dir(self, *paths):
+        return get_bedrock_gdk("Minecraft Bedrock", *paths) or self._get_game_uwp(
+            *paths
+        )
 
     def _get_logs_dir(self):
         game_dir = self.get_game_dir()