+++ mcpath/platforms/win/preview.py
@@ -5,24 +5,13 @@ Windows Preview Edition
 from typing import Optional
 from os import path
 from mcpath.facades import Preview
-from mcpath.utils import step_back
+from mcpath.utils import step_back, get_bedrock_gdk
 
 
 class WinPreviewEdition(Preview):
 
     def is_old(self, path: str) -> bool:
         return "Microsoft.MinecraftWindowsBeta" in path
-
-    def _get_game_gdk(self) -> Optional[str]:
-        """
-        New path
-        """
-        p = path.expandvars(
-            "%appdata%\\Minecraft Bedrock Preview\\users\\shared\\games\\com.mojang"
-        )
-        if path.isdir(p):
-            return p
-        return None
 
     def _get_game_uwp(self) -> Optional[str]:
         """
@@ -35,13 +24,10 @@ class WinPreviewEdition(Preview):
             return p
         return None
 
-    def _get_game_dir(self):
-        gdk = self._get_game_gdk()
-        if gdk:
-            return gdk
-
-        # Fallback to old path
-        return self._get_game_uwp()
+    def _get_game_dir(self, *paths: str):
+        return get_bedrock_gdk(
+            "Minecraft Bedrock Preview", *paths
+        ) or self._get_game_uwp(*paths)
 
     def _get_executable(self):
         return "minecraft-preview://"