+++ tests/test_java.py
@@ -14,6 +14,11 @@ def test_java():
     mcpath.java.get_resource_packs_dir()
     mcpath.java.get_screenshots_dir()
     mcpath.java.get_logs_dir()
+    mcpath.java.get_versions()
+    mcpath.java.get_saves()
+    mcpath.java.get_resource_packs()
+    mcpath.java.get_screenshots()
+    mcpath.java.get_logs()
 
 
 def test_deprecated_java():