+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - 11/1/2025
+
+### General
+
+- Added functions to list worlds, packs, screenshots, logs and more.
+
+### Fixes
+
+- Bedrock and preview return an invalid world GDK path.
+
 ## [2.0.2] - 10/28/2025
 
 ### General